moved socket_service.py to services
