+++ src/routers/ws_routers.py
@@ -1,5 +1,3 @@
-import json
-
 from fastapi import APIRouter, Depends, WebSocket, WebSocketDisconnect
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -8,7 +6,7 @@ from db.database import get_db
 from dependencies import get_current_user
 from repositories.user_room_repository import UserRoomRepository
 from schemas.user_room_schemas import UserRoomResponseSchema
-from websocket.socket_service import ConnectionManager
+from services.socket_service import ConnectionManager
 
 
 router = APIRouter()
